extractor: Use correct size and sign dealing with read() in mp3 extractor
When reading ID3v1 tags, use the right sign to check for read() returning
errors.

CID: #457929

+++ src/extractor/tracker-extract-mp3.c
@@ -631,8 +631,8 @@ read_id3v1_buffer (int     fd,
                    goffset size)
 {
 	char *buffer;
-	guint bytes_read;
-	guint rc;
+	size_t bytes_read;
+	ssize_t rc;
 
 	if (size < 128) {
 		return NULL;
